Add new replay for location.
Add a function to convert the time and date from uint32_t to a char*
Tried to replace the needle by sprites, failed so far.

Create a new function displaying a replay displaying the track dot per dot, using the longitude and latitude.

Corrected a bug in the sd function, add the sorting of the files.