Changed the main data structure
To make it easier to communicate with the android app, the data_point struct is modified, it is now packed and take only the space really used without padding.
Also, the speed and direction were changed from double to float to save space.
We moved from 56bits per point to 44, so about 20% reduction.