Add replay scaling and borders
In the replay, the data are not scaled to best fit the replay area.
The lat and lng will be inverted if needed to fit the screen better. This is not tested yet since I do not have data to test that.
The replay also has some small border (4% on each side) to look nicer.