Improve app and soft
For the M5, reduce the wait time, improve writting on the sd card.
Now about 8.1 pt per second for a 1hours recording time.
When the files grow, the number of  pt per second decrease a bit.

For the app,
Added a rotating background for the recap.
When reading the file, interpolate the time for the point having the same second, this improve the precision for laptime.