Faster lap time computation. Empty lap removed
New function to compute the distance between two points. Since we are on track, we can assume that earth is flat on these kind of distances. So we change the complex computation by a simple flat triangle.

To estimate the laptime, we were waiting 20 seconds. Does not really work for the beginning and end of the capture. So we change the wait 20 seconds by be at least 100m from the starting point.