Improve transfer speed, add download info
For whatever reason, the M5 can transfer several bytes at a time, but the android has to receive them one by one. Since the M5 was the bottleneck, I now transfer per packet of 44 bytes (one data point), so it is much much faster.
Also, when we transfer, we display it on the M5 screen and the android app, to show that something is happeneing.